+++ prototype/drivers/hm55b.h
@@ -1,0 +1,14 @@
+/*
+ * Sierra College Computer Engineering Club: Sparkfun AVC
+ * hm55b magnetometer driver code
+ * Authors: 
+ */
+
+#ifndef _MAGNETOMETER_H_
+#define _MAGNETOMETER_H_
+
+vector mag_reset();
+vector mag_read();
+
+#endif
+
